requirements-web.txt
# Web review interface requirements
Flask>=3.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
